Search / Filter Optimization
1. Made tables for posts more DB normalized. You don't want to filter on JSON DB data, it should only ever be used for presentation data.
2. Pushed all migrations to app config, so you can choose How your table should behave. 
3. Added steps to post repository to merge updated data with an empty array of fields to that fields left out are overwritten with null values.
4. Post repository now posts Local Field Type as well as built in types
For example you could now have a App\Fields\Series and App\Field\Parts and they would be saved to the post table if you added them in the migration. This will keep them out of the extra field as they are probably something you might want to run DB queries on.